+++ packages/a11y-speak/src/clear.js
@@ -1,0 +1,11 @@
+/**
+ * Clear the live regions.
+ */
+const clear = function() {
+	let regions = document.querySelectorAll( ".a11y-speak-region" );
+	for ( let i = 0; i < regions.length; i++ ) {
+		regions[ i ].textContent = "";
+	}
+};
+
+export default clear;
